Conclusione della separazione dei metodi in classe DbAndBusiness
