+++ BusinessLayer/AllStudents.cs
@@ -1,0 +1,54 @@
+﻿using System;
+using System.Data.Common;
+using System.Data;
+using System.Data.SQLite;
+using System.IO;
+using System.Collections.Generic;
+using gamon;
+using System.Windows.Forms;
+using SchoolGrades.DbClasses;
+
+namespace SchoolGrades.BusinessLayer
+{
+    class AllStudents
+    {
+
+        internal void BackupAllStudentsDataTsv()
+        {
+            BackupTableTsv("Students");
+            BackupTableTsv("StudentsPhotos");
+            BackupTableTsv("StudentsPhotos_Students");
+            BackupTableTsv("Classes_Students");
+            BackupTableTsv("Grades");
+        }
+
+        //-----------------RIPARTIRE DAL DI QUA-----------------
+
+        internal void BackupAllStudentsDataXml()
+        {
+            BackupTableXml("Students");
+            BackupTableXml("StudentsPhotos");
+            BackupTableXml("StudentsPhotos_Students");
+            BackupTableXml("Classes_Students");
+            BackupTableXml("Grades");
+        }
+
+        internal void RestoreAllStudentsDataTsv(bool MustErase)
+        {
+            RestoreTableTsv("Students", MustErase);
+            RestoreTableTsv("StudentsPhotos", MustErase);
+            RestoreTableTsv("StudentsPhotos_Students", MustErase);
+            RestoreTableTsv("Classes_Students", MustErase);
+            RestoreTableTsv("Grades", MustErase);
+        }
+
+        internal void RestoreAllStudentsDataXml(bool MustErase)
+        {
+            RestoreTableXml("Students", MustErase);
+            RestoreTableXml("StudentsPhotos", MustErase);
+            RestoreTableXml("StudentsPhotos_Students", MustErase);
+            RestoreTableXml("Classes_Students", MustErase);
+            RestoreTableXml("Grades", MustErase);
+        }
+    }
+}
