+++ DataLayer/Class.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SchoolGrades.DataLayer
+{
+    class Class
+    {
+
+    }
+}
