--- SchoolGrades/bin/Release/netcoreapp3.1/SchoolGrades.deps.json
@@ -1,75 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v3.1",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v3.1": {
-      "SchoolGrades/0.51.0.1": {
-        "dependencies": {
-          "System.Data.SQLite.Core": "1.0.113.7"
-        },
-        "runtime": {
-          "SchoolGrades.dll": {}
-        }
-      },
-      "Stub.System.Data.SQLite.Core.NetStandard/1.0.113.2": {
-        "runtime": {
-          "lib/netstandard2.1/System.Data.SQLite.dll": {
-            "assemblyVersion": "1.0.113.0",
-            "fileVersion": "1.0.113.0"
-          }
-        },
-        "runtimeTargets": {
-          "runtimes/linux-x64/native/SQLite.Interop.dll": {
-            "rid": "linux-x64",
-            "assetType": "native",
-            "fileVersion": "0.0.0.0"
-          },
-          "runtimes/osx-x64/native/SQLite.Interop.dll": {
-            "rid": "osx-x64",
-            "assetType": "native",
-            "fileVersion": "0.0.0.0"
-          },
-          "runtimes/win-x64/native/SQLite.Interop.dll": {
-            "rid": "win-x64",
-            "assetType": "native",
-            "fileVersion": "1.0.113.0"
-          },
-          "runtimes/win-x86/native/SQLite.Interop.dll": {
-            "rid": "win-x86",
-            "assetType": "native",
-            "fileVersion": "1.0.113.0"
-          }
-        }
-      },
-      "System.Data.SQLite.Core/1.0.113.7": {
-        "dependencies": {
-          "Stub.System.Data.SQLite.Core.NetStandard": "1.0.113.2"
-        }
-      }
-    }
-  },
-  "libraries": {
-    "SchoolGrades/0.51.0.1": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    },
-    "Stub.System.Data.SQLite.Core.NetStandard/1.0.113.2": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-yGIM9xFyPHl79HluNTSQarvAOIJ68pmzmlMayVN6Nivu/7TiaHt7WqptolvwmuKdAXei8OzENh9RNdh7d6evjQ==",
-      "path": "stub.system.data.sqlite.core.netstandard/1.0.113.2",
-      "hashPath": "stub.system.data.sqlite.core.netstandard.1.0.113.2.nupkg.sha512"
-    },
-    "System.Data.SQLite.Core/1.0.113.7": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-2DKMIxfZpIVzdMfx1dj9qvTm+0m0Mx6sN6CZXMQDNytU7Sk/JKo0Ox6mvHAicfHrQgqZbUZMzMMtLXD0kwmFfg==",
-      "path": "system.data.sqlite.core/1.0.113.7",
-      "hashPath": "system.data.sqlite.core.1.0.113.7.nupkg.sha512"
-    }
-  }
-}
