--- SchoolGrades/bin/Release/netcoreapp3.1/SchoolGrades.runtimeconfig.json
@@ -1,9 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "netcoreapp3.1",
-    "framework": {
-      "name": "Microsoft.WindowsDesktop.App",
-      "version": "3.1.0"
-    }
-  }
-}
